chore: Optimize script log output
